updated code

+++ app/src/main/java/com/vital_self/features/history/data/model/SaveScanHistoryModels.kt
@@ -73,7 +73,7 @@ data class SaveScanHistoryRequest(
     @SerializedName("heartAge")
     val heartAge: Int?,
     @SerializedName("ascvdRiskLevel")
-    val ascvdRiskLevel: Int?,
+    val ascvdRiskLevel: Double?,
     @SerializedName("cardiacWorkload")
     val cardiacWorkload: Double?,
     @SerializedName("pulsePressure")
